Move tests to Orchestration namespace
